Adjusted examples to Amaranth 0.4.0+
In Amaranth 0.4.0 the Pin instance can no longer simply be cast to a
Value, instead you have to specify the input/output direction when
accessing it. This commit makes these changes for the blink and uart
example of the icebreaker. Changes were not made to the other examples
as I can't test them here.