Adjusted examples to Amaranth 0.4.0+ (UNTESTED)
In Amaranth 0.4.0 the Pin instance can no longer simply be cast to a
Value, instead you have to specify the input/output direction when
accessing it. This commit makes these changes for the remaining
examples of the icebreaker. The changes were flashed to the icebreaker
but I'm unable to test them as I don't have the accessories. The
pdm_fade_gamma example unfortunately doesn't build for me at all so I
didn't fix that.